add base package concept

+++ misc/devenv/li-devenv.sh
@@ -121,13 +121,49 @@ yum -y install java-1.6.0-openjdk.i686 java-1.6.0-openjdk-devel.i686
 
 function install_build_requires {
   spec_file=$1
+  requires=""
   for s in $(grep -e "^BuildRequires:" $spec_file)
   do
-    if [[ $s =~ ^[a-z]+ ]]
+    if [[ $s != "BuildRequires:" ]] && [[ $s =~ ^[A-Z|a-z]+ ]]
     then
-      yum install -y $s
+      requires+=" $s"
     fi
   done
+  if [ -n "$requires" ]
+  then
+    yum install -y $requires
+  fi
+}
+
+function install_requires {
+  spec_file=$1
+  requires=""
+  output=$1
+  IFS_BAK=$IFS
+IFS="
+"
+  for s in $(grep -e "^Requires:" $spec_file)
+  do
+    if ! [[ $s =~ "%{" ]]
+    then
+      s=`echo ${s//Requires:/}`
+      s=`echo ${s// /}`
+      s=`echo ${s//(/-}`
+      s=`echo ${s//)/}`
+      s=`echo ${s//>=*/}`
+      s=`echo ${s//=/-}`
+      s=`echo ${s//,/ }`
+      if ! `contains_value "$s" "${@:2}"`
+      then
+        requires+=" $s"
+      fi
+    fi
+  done
+  IFS=$IFS_BAK
+  if [ -n "$requires" ]
+  then
+    yum install -y $requires
+  fi
 }
 
 function find_and_build_specs {
@@ -144,35 +180,49 @@ function find_and_build_specs {
   done
 }
 
+function contains_value { 
+  for v in "${@:2}"
+  do 
+    [[ "$v" = "$1" ]] && break
+  done
+}
+
+github_repos=( crankcase os-client-tools )
+repos=( crankcase os-client-tools li-working )
+
+branch="master"
+if [ "$1" == "stage" ]
+then
+  branch="stage"
+fi
+
+rm -rf /root/li-working
+
+if [[ "$2" == "--install_from_source" ]] || [[ "$2" == "--install_required_packages" ]]
+then
+  git clone git://git1.ops.rhcloud.com/li.git/ /root/li-working
+
+  for repo_name in "${github_repos[@]}"
+  do
+    rm -rf /root/$repo_name
+    git clone git@github.com:openshift/$repo_name.git /root/$repo_name
+  done
+elif [[ "$2" == "--install_from_local_source" ]]
+then
+  git clone /root/li /root/li-working
+fi
+
 if [[ "$2" == "--install_from_source" ]] || [[ "$2" == "--install_from_local_source" ]]
 then
+
   mkdir -p /tmp/tito
-  rm -rf /root/li-working
-  if [[ "$2" == "--install_from_source" ]]
-  then
-    git clone git://git1.ops.rhcloud.com/li.git/ /root/li-working
-  else
-    git clone /root/li /root/li-working
-  fi
 
-  github_repos=( crankcase os-client-tools )
-
-  if ! [[ "$2" == "--install_from_local_source" ]]
-  then
-    for repo_name in "${github_repos[@]}"
-    do
-      rm -rf /root/$repo_name
-      git clone git@github.com:openshift/$repo_name.git /root/$repo_name
-    done
-  fi
-
-  pushd /root/li-working > /dev/null
-  find_and_build_specs
-  for repo_name in "${github_repos[@]}"
+  for repo_name in "${repos[@]}"
   do
     if [ -d /root/$repo_name ]
     then
       pushd /root/$repo_name > /dev/null
+        git checkout $branch
         find_and_build_specs
       popd > /dev/null
     fi
@@ -194,22 +244,59 @@ EOF
   createrepo /root/li-local/
 
   yum -y install rhc-devenv --enablerepo=li-local
-
-  build/devenv write_sync_history
   
+  pushd /root/li-working > /dev/null
+    build/devenv write_sync_history
   popd > /dev/null
-  for repo_name in "${github_repos[@]}"
+elif [[ "$2" == "--install_required_packages" ]]
+then
+  pushd /root/li-working > /dev/null
+    git checkout $branch
+    packages_str=`build/devenv print_packages`
+    packages=($packages_str)
+    ignore_packages_str=`build/devenv print_ignore_packages`
+    ignore_packages=($ignore_packages_str)
+  popd > /dev/null
+  for repo_name in "${repos[@]}"
   do
-    rm -rf /root/$repo_name
-  done
-  rm -rf /root/li-working /tmp/tito
-  for repo_name in "${github_repos[@]}"
-  do
-    git init --bare /root/$repo_name  
+    pushd /root/$repo_name > /dev/null
+      git checkout $branch
+	  for x in $(find -name *.spec)
+	  do
+	    name_grep=`grep -e "^Name:" $x`
+	    name_grep_array=($name_grep)
+	    package_name=${name_grep_array[1]}
+        if [[ $package_name =~ "%{gemname}" ]]
+        then
+          gemname_grep=`grep -e "%global gemname" $x`
+          gemname_grep_array=($gemname_grep)
+          gemname=${gemname_grep_array[2]}
+          package_name=`echo ${package_name//%\{gemname\}/$gemname}`
+        fi
+        if ! `contains_value "$package_name" "${ignore_packages[@]}"`
+	    then
+	      install_build_requires "$x"
+	      install_requires "$x" "${packages[@]}"
+	    fi
+	  done
+	popd > /dev/null
   done
 elif [ "$2" == "--install_build_prereqs" ]
 then
   yum -y install git tito
 else
   yum -y install rhc-devenv
+fi
+
+if [[ "$2" == "--install_from_source" ]] || [[ "$2" == "--install_from_local_source" ]] || [[ "$2" == "--install_required_packages" ]]
+then
+  rm -rf /root/li-working /tmp/tito
+  for repo_name in "${github_repos[@]}"
+  do
+    rm -rf /root/$repo_name
+    if ! [[ "$2" == "--install_required_packages" ]]
+    then
+      git init --bare /root/$repo_name
+    fi
+  done
 fi
